Move the delete account questionnaire to separate component
